% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acccess_fire_hazard.R
\name{access_fire_hazard}
\alias{access_fire_hazard}
\title{Access fire hazard data store locally or remotely}
\usage{
access_fire_hazard(
  data_path,
  file_pattern = NULL,
  aoi = NULL,
  output_path = NULL
)
}
\arguments{
\item{data_path}{Local or remote path or URL to gridded data files describing fire hazard}

\item{file_pattern}{By default, the function expects "\emph{_bp.tif" for Burn Probability and "}_nvc_flp_X_X.tif" for flame length probabilities
Can optionally provide alternative patterns as a vector of strings, like \code{c("pattern1.tif", "pattern2.tif")}}

\item{aoi}{Optionally provide an Area Of Interest, which the hazard data will be cropped and masked to.}

\item{output_path}{Path to save outputs. Defaults holding in memory.}
}
\value{
Vector of character strings file paths to the cropped files
}
\description{
Access fire hazard data store locally or remotely
}
